+++ idealingua/idealingua-runtime-rpc-scala/src/main/scala/com/github/pshirshov/izumi/idealingua/runtime/rpc/IRTMethodSignature.scala
@@ -1,0 +1,8 @@
+package com.github.pshirshov.izumi.idealingua.runtime.rpc
+
+trait IRTMethodSignature extends IRTZioResult {
+  type Input <: Product
+  type Output <: Product
+
+  def id: IRTMethodId
+}
