--- idealingua/idealingua-runtime-rpc-scala/src/main/scala/com/github/pshirshov/izumi/idealingua/runtime/rpc/IRTRawCall.scala
@@ -1,5 +1,0 @@
-package com.github.pshirshov.izumi.idealingua.runtime.rpc
-
-import io.circe.Json
-
-final case class IRTRawCall(methodId: IRTMethodId, body: Json)
