+++ idealingua/idealingua-runtime-rpc-scala/src/main/scala/com/github/pshirshov/izumi/idealingua/runtime/rpc/models.scala
@@ -1,0 +1,34 @@
+package com.github.pshirshov.izumi.idealingua.runtime.rpc
+
+import io.circe.Json
+
+
+// addressing
+final case class IRTServiceId(value: String) extends AnyVal {
+  override def toString: String = s"{service:$value}"
+}
+
+final case class IRTMethodName(value: String) extends AnyVal {
+  override def toString: String = s"{method:$value}"
+}
+
+final case class IRTMethodId(service: IRTServiceId, methodId: IRTMethodName) {
+  override def toString: String = s"${service.value}.${methodId.value}"
+}
+
+
+// dtos
+
+final case class IRTReqBody(value: Product) extends AnyRef
+
+final case class IRTResBody(value: Product) extends AnyRef
+
+final case class IRTMuxResponse[T <: Product](v: T, method: IRTMethodId) {
+  def body: IRTResBody = IRTResBody(v)
+}
+
+final case class IRTMuxRequest[T <: Product](v: T, method: IRTMethodId) {
+  def body: IRTReqBody = IRTReqBody(v)
+}
+
+final case class IRTJsonBody(methodId: IRTMethodId, body: Json)
